bug fix: editing mode having different text
There was a bug where after deleting the completed tasks, if we were editing any task, it's state (value of edit field) and prop (value of task title) were not being in sync, hence causing the bug.

With useEffect hook, updating the local state for task whenever there is a change in prop.
This fixes the issue.